Use BACKEND status of HAproxy stats for available check.
Instead of counting backend nodes, which are UP, we can get the
aggregate status by parsing the BACKEND entry for the cluster in the
HAproxy stats output. This resolves a subtle bug where nodes return "UP
x/y" when they are going down and x out of y HAproxy health checks have
failed, and this output was not properly parsed as "available".