Add the ability to force a specific health measurement
Previously, if you wanted to force a specific health measurement for
a service, you needed to change the configuration for that service to
include a constant metric at the value you wanted. This commit adds
(albeit a bit hackishly) the ability to specity an exact health level:

  litmusctl force health 88 -r "Because reasons"

...would then set the measured health of the service to 88.

This is sometimes useful if you want to slowly balance in/out a server,
and you want to do that via litmus.