Use measured_health instead of forced health value for unhealthy things
After thinking about things for awhile, I realized that you probably
don't want to direct traffic to a down service just because the health
value is forced to something non-zero. In reality, if service is failing
and you want to force traffic there, you should probably force it up
until you fix the problem, then remove the up-file and force it back to
the desired health.

Otherwise, you won't be able to detect a truly failing service when it
has its health forced for traffic-diverting reasons.