Add internet health metric
This metric provides a simple of heuristic of how healthy an ISP
connection appears to be. You provide it a list of host and ports:

LitmusPaper::Metric::InternetHealth.new([
  "cloud.google.com:443",
  "azure.microsoft.com:443",
  "aws.amazon.com:443",
])

And the check performs a TCP connect to each host and port. The metric
then returns a number between 0 and 100 indicating the percentage of
host which are reachable.