Add HAProxy agent-check support
This commit adds a new program litmus-agent-check which may be used as
an agent check for HAProxy. This program would need to be run under an
inetd type server. Given that litmus runs its checks inline with its
call this method provides a simple way to add agent-check support. An
alternative would be to add a TCP server to litmus, but that has few
advantages over a well tested server such as xinetd. The litmus states
are mirrored with as much fidelity as possible to the agent-check
states.

Example output:

  ready   100%

Example config for xinetd:

  service internet_health
  {
    disable     = no
    socket_type = stream
    protocol    = tcp
    port        = 9191
    user        = agent-check
    wait        = no
    server      = /usr/bin/litmus-agent-check
    server_args = -s isp
    type        = UNLISTED
    instances   = 10
  }

HAProxy agent-check documentation:

https://cbonte.github.io/haproxy-dconv/1.6/configuration.html#5.2-agent-check