Symbolize keys in YAML configs to match eval'ed behavior
This brings YAML configuration to parity with eval'ed configs, save
for using ruby to dynamically generate config. The tests are a little
unfortunate (#instance_variable_get) but in this pass, I am erring on
the side of avoiding as many changes to existing code as possible, and
the current classes expose next to nothing via conventional means.