Fix agent-check to correctly set operational state to up
HAProxy has two states for a server, its operational state and its
administrative state. Prior to this commit when a server is up, we were
only setting the administrative state to `ready`. This however is
insufficient if the operating state was previously `down`. In order to
cancel a `down` operational state the agent check must report `up`.

This commit sets the operational state to `up` whenever we set the
administrative state to `ready`