Set per host timeout as a portion of configured timeout for internet_health
Prior to this commit the configuration option :timeout_seconds set the
timeout value for each connection attempt not the timeout for the entire
internet health metric. This made its usage confusing as you would need
to multiply the configured timeout by the number of hosts to obtain the
timeout for the entire metric.

This change divides the configured timeout by the number of hosts, so
that the entire metric returns shortly after the configured timeout.