Add simple Vagrant setup
It's useful to be able to get a VM with dependencies built and installed,
and also to play around with the service.

The worst part of all of this is that things like `bin/litmus` and
`bin/litmusctl` expect the litmus gem to be installed (and hence we
build/install it in the vm). Future work should make this run
from the checkout w/o actually installing the gem.