Add active and queued requests to text output of unix socket metric
The UNIX socket metric makes use of active and queued socket connections
as part of calculating the final weight. Those active and queued metrics
can be valuable as they provide insight into the number of received but
not-yet processed requests. By exposing those metrics in the text output
detail, other utilities can extract that information from outside the
host running the service. This can be especially useful when the service
is running in a container, it allows services outside the container such
as monitoring/metrics agents to collect this information without needing
direct access to the host or container itself.

These metrics are also valuable for application of some queuing theory
formulas, such as Little's Law. Active + Queued requests provide the L
in Little's formula. And since the metric also exposes the maximum
number of connections, this allows a third-party service such as a
monitoring agent to use (active + queued) / max to calculate how close
to maximum utilization the service running in the container is.